Help: Moving all the help menu related stuff to a new super class
